使用正面投影的方法计算行人数量
1. 首先，将行人通过时，被检测到的突起利用正面投影的方法累积起来，减少突起宽度的限制。如此，可以从最后累积的结果看到比较完整的头肩形状
2. 计算行人的方向依旧是取最大值点，然后判断最值点左右之点数之多少
3. 在判断相邻两次检测的突起是否可以归并时，修改了 operator-重载函数，令其返回重合的监测点至于两个突起的重复度
4. 当前之代码仅应用与 Mar 的测试数据。除了第一个测试集，其它结果皆有问题。
5. 此份代码还有突起对其的问题尚未处理（应对行人斜行）

+++ Prominent.h
@@ -32,8 +32,9 @@ public:
 
 
 
-	// overload the operator -, which calculates the difference between tow segments
-	int operator-(const Prominent &r) const;
+	// overload the operator -, which calculates the matching rate
+	// between two segments, say the number of common detection points
+	double operator-(const Prominent &r) const;
 
 
 	int start;
